requirements.txt
# Web scraping
requests
beautifulsoup4
tqdm

# Data I/O e processamento
pyarrow
s3fs
jinja2 
joblib
argparse
loguru
python-dotenv
pydantic

# Grafos e visualização
networkx
pygraphviz
plotly

# ML & Otimização
scipy
xgboost
optuna
scikit-learn
dask-ml

# Dask e dependências específicas dos workers
dask[distributed,dataframe]
bokeh>=3.1.0

mlflow
loguru
boto3
setuptools
lightgbm
python-dotenv

joblib
optuna
imblearn

#Henrique
# pip install dask[complete] dask-ml dask-cuda distributed \
# xgboost optuna scikit-learn matplotlib \
# numpy pandas zarr h5py joblib torch